% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{extract_category_pages}
\alias{extract_category_pages}
\title{Get all page URLs belonging to a Category entry}
\usage{
extract_category_pages(home_page)
}
\arguments{
\item{home_page}{landing page url of a Category}
}
\value{
character vector
}
\description{
Get all page URLs belonging to a Category entry
}
\examples{
extract_category_pages("https://www.mixesdb.com/w/Category:Essential_Mix")
}
