% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixes.R
\name{get_mix_meta}
\alias{get_mix_meta}
\title{Get mix-level metadata}
\usage{
get_mix_meta(url)
}
\arguments{
\item{url}{URL of a mix page}
}
\value{
list
}
\description{
Get mix-level metadata
}
\examples{
\dontrun{
mixes <- get_mix_entries(url = "https://www.mixesdb.com/w/Category:FACT_Mix")
get_mix_meta(mixes[1:5])
}
}
