% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixes.R
\name{get_mix_meta_tidy}
\alias{get_mix_meta_tidy}
\title{Get mix-level metadata into a tidy data frame}
\usage{
get_mix_meta_tidy(url)
}
\arguments{
\item{url}{URL of a mix page}
}
\value{
tibble
}
\description{
Get mix-level metadata into a tidy data frame
}
\examples{
\dontrun{
mixes <- get_mix_entries(url = "https://www.mixesdb.com/w/Category:FACT_Mix")
get_mix_meta_tidy(mixes[1:5])
}
}
