% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracklist.R
\name{get_tracklist}
\alias{get_tracklist}
\title{Read tracklist metadata from a mix entry}
\usage{
get_tracklist(url)
}
\arguments{
\item{url}{URL of a mix entry}
}
\value{
a list
}
\description{
Read tracklist metadata from a mix entry
}
\examples{
\dontrun{
get_tracklist(url = "https://www.mixesdb.com/w/2019-03-06_-_Object_Blue_-_The_Selector_After_Dark")
get_tracklist(url = "https://www.mixesdb.com/w/2020-06-15_-_Kush_Jones_-_Resident_Advisor_(RA.732)")
}
}
