% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracklist.R
\name{get_tracklist_tidy}
\alias{get_tracklist_tidy}
\title{Get tracklist metadata from a mix entry into a tidy data frame}
\usage{
get_tracklist_tidy(url)
}
\arguments{
\item{url}{URL of a mix page}
}
\value{
a tibble
}
\description{
Get tracklist metadata from a mix entry into a tidy data frame
}
\examples{
\dontrun{
get_tracklist_tidy("https://www.mixesdb.com/w/2011-08-23_-_Objekt_@_Boiler_Room_Berlin_001")
}
}
