% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{read_category_entries}
\alias{read_category_entries}
\title{Get all mix URLs listed under a Category entry}
\usage{
read_category_entries(home_page)
}
\arguments{
\item{home_page}{landing page url of a Category}
}
\value{
character vector
}
\description{
Get all mix URLs listed under a Category entry
}
\examples{
read_category_entries("https://www.mixesdb.com/w/Category:Essential_Mix")
}
