% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracklist.R
\name{read_tracklist}
\alias{read_tracklist}
\title{Read tracklist metadata from a mix entry}
\usage{
read_tracklist(page)
}
\arguments{
\item{page}{url of a mix entry}
}
\value{
tibble
}
\description{
Read tracklist metadata from a mix entry
}
\examples{
read_tracklist("https://www.mixesdb.com/w/2011-08-23_-_Objekt_@_Boiler_Room_Berlin_001")
}
