% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{get_mix_entries}
\alias{get_mix_entries}
\title{Get all mix entries listed under a sub-category}
\usage{
get_mix_entries(home_page)
}
\arguments{
\item{home_page}{home page URL of a sub-category}
}
\value{
character vector
}
\description{
Get all mix entries listed under a sub-category
}
\examples{
get_mix_entries(home_page = "https://www.mixesdb.com/w/Category:FACT_Mix")
get_mix_entries(home_page = "https://www.mixesdb.com/w/Category:fabric")
get_mix_entries(home_page = "https://www.mixesdb.com/w/Category:Ben_UFO")
}
