% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{get_subcat_entries}
\alias{get_subcat_entries}
\title{Get all sub-category entries listed on a category page}
\usage{
get_subcat_entries(home_page)
}
\arguments{
\item{home_page}{home page URL of a category}
}
\value{
character vector
}
\description{
Get all sub-category entries listed on a category page
}
\examples{
get_subcat_entries(home_page = "https://www.mixesdb.com/w/MixesDB:Regular_categories")
get_subcat_entries(home_page = "https://www.mixesdb.com/w/Category:Venue")
}
